requirements.txt
certifi==2020.11.8
email-validator==1.1.2
fastapi==0.61.2
fastapi-sqlalchemy==0.2.1
opennmt-tf==2.9.3
passlib==1.7.4
pip-chill==1.0.0
pyodbc==4.0.30
python-dotenv==0.15.0
tenacity==6.2.0
uvicorn==0.12.3
wincertstore==0.2
